add footerText
footerText was added upstream (in the BrownCLPS/LingView repo), and we've run the Prepare Upgrade script since then, but the Prepare Upgrade script ignores all changes to LocaleConstants.jsx file. This caused the site to break (blank white screen, "Uncaught TypeError: dictionary is not defined"). We need to manually add the definition of footerText to fix it.